requirements.txt
# AI
langchain
langchain-core
langchain-openai
langchain-community
langsmith
openai
# Tools
black
faiss-cpu
python-dotenv
docx2txt
pypdf
# UI
streamlit
